requirements.txt
dash
plotly
pandas
pyarrow
numpy
dask
pymrio
python-dotenv
PyYAML
dash-bootstrap-components
pytest
scipy